requirements.txt
alembic==1.8.1
ariadne==0.16.0
Flask==2.2.2
Flask-Migrate==3.1.0
Flask-Sqlalchemy==2.5.1
Flask-Minio==0.1.2
PyJWT==2.4.0
python-dotenv==0.21.0
SQLAlchemy==1.4.41
Werkzeug==2.2.2
